# openapi.yaml
openapi: 3.0.3
info:
  title: API Documentation
  version: '1.0'
servers:
  - url: http://localhost:8080
paths:
...
  /ads:
    get:
      tags:
        - Объявления
      operationId: getAllAds
      summary: 'Получение всех объявлений'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ads'
    post:
      tags:
        - Объявления
      summary: 'Добавление объявления'
      operationId: addAd
      requestBody:
        content:
          multipart/form-data:
            schema:
              required:
                - image
                - properties
              type: object
              properties:
                properties:
                  $ref: '#/components/schemas/CreateOrUpdateAd'
                image:
                  type: string
                  format: binary
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdDto'
        '401':
          description: Unauthorized
...
  /ads/{id}:
    get:
      tags:
        - Объявления
      summary: 'Получение информации об объявлении'
      operationId: getAds
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExtendedAdDto'
        '401':
          description: Unauthorized
        '404':
          description: Not found
    delete:
      tags:
        - Объявления
      summary: 'Удаление объявления'
      operationId: removeAd
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
    patch:
      tags:
        - Объявления
      summary: 'Обновление информации об объявлении'
      operationId: updateAds
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrUpdateAd'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdDto'
        '403':
          description: Forbidden
        '401':
          description: Unauthorized
        '404':
          description: Not found
...
  /ads/me:
    get:
      tags:
        - Объявления
      summary: 'Получение объявлений авторизованного пользователя'
      operationId: getAdsMe
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ads'
          description: OK
        '401':
          description: Unauthorized
...
components:
  schemas:
...
    CreateOrUpdateAd:
      type: object
      properties:
        title:
          type: string
          description: 'заголовок объявления'
          minLength: 4
          maxLength: 32
        price:
          type: integer
          format: int32
          description: 'цена объявления'
          minimum: 0
          maximum: 10000000
        description:
          type: string
          description: 'описание объявления'
          minLength: 8
          maxLength: 64
...
    Ads:
      type: object
      properties:
        count:
          type: integer
          format: int32
          description: 'общее количество объявлений'
        results:
          type: array
          items:
            $ref: '#/components/schemas/AdDto'
...